Refactor SSJ codebase a bit more
"Session" has been renamed to "inferior", aligning SSJ more with the GDB
concepts it was based on.  Next thing I want to do is move the
command-line handling out of inferior.c since it doesn't really belong
there.

+++ src/ssj/sockets.h
@@ -7,7 +7,7 @@ void      sockets_init   (void);
 void      sockets_deinit (void);
 socket_t* socket_connect (const char* hostname, int port, double timeout);
 void      socket_close   (socket_t* socket);
-size_t    socket_recv    (socket_t* socket, void* buffer, size_t num_bytes);
-size_t    socket_send    (socket_t* socket, const void* data, size_t num_bytes);
+int       socket_recv    (socket_t* socket, void* buffer, int num_bytes);
+int       socket_send    (socket_t* socket, const void* data, int num_bytes);
 
 #endif // SSJ__SOCKET_H__INCLUDED
